pyproject.toml
[project]
name = "chatter-pipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chatterbox-tts>=0.1.1",
    "sounddevice>=0.5.2",
]
